requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
mysqlclient==2.0.3
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
